+++ src/misc/load_kmeans.hh
@@ -3,6 +3,6 @@
 #include <string>
 #include <vector>
 
-using centroid_type = double;
+using centroid_type = float;
 
 std::vector<centroid_type> load_kmean_centroids(const std::string& filepath);
